Validate Vulkan API version when creating device.
This CL adds validation into the layer CreateDevice method to validate
that the api_version selected is at least the same as the api_version
requested in vkInstanceCreate.

If the device version is lower, this can cause issues in later layers
(like shader_validation) where the lower version number is used and it
isn't obvious what happened.

Fixes #951