layers: Refine query state tracking and errors
This commit removes query-state-related errors generated from calls to
vkGetQueryPoolResults(). The reasoning is that this function operates
asynchronously to work done on the GPU.

For example, if a query is reset from the host using
vkResetQueryPoolEXT(), then work is submitted to a queue beginning and
ending this query, then vkGetQueryPoolResults() is called using the wait
bit without waiting for the queue or device to be idle, the validation
layers would emit an error indicating the code is waiting on a query
that has been reset and not issued. This error is spurious because what
would happen is that the call to vkGetQueryPoolResults() succeeds after
waiting until the GPU finishes the query and makes results available for
it.

Similar circumstances surround other types of errors related to query
states and vkGetQueryPoolResults(). It's possible to get spurious errors
for any of them. Some of these happen in a subset of VK-GL-CTS tests
when enabling validation.

However, this commit does *not* remove similar errors for
vkCmdCopyQueryPoolResults(), which is submitted to command buffers and
does not have this kind of problems.

In addition, this commit adds one more state to queries, allowing the
validation layers to make a distinction between ended queries and
available queries. Queries are switched from the ended state to the
available state when retiring work on queues after waiting for the queue
or device to be idle, or when waiting on a fence.

This refinement fixes more spurious errors in some VK-GL-CTS tests.

Change-Id: Ia7055a3145031d84a4efac8031c851549c679ebb