layers: Hack to remove incorrect implicit VU
VUID-VkFramebufferCreateInfo-pAttachments-parameter should've been
removed for VK_KHR_imageless_framebuffer, and replaced with explicit
VU instead - but this hasn't yet happened.
In order to ship validation for this functionality in the next SDK,
the VU has temporarily been hardcoded out in the generator script.
Added in an explicit check temporarily to ensure that

Most of these changes (other than the change to
parameter_validation_generator.py) will remain intact after the spec is
updated. The script change won't cause any issues after it the spec is
updated either, so it can be reverted in slow time.