layers: Use persistent objects for ThreadSafety tracking
The previous 'Bucket' change was a big improvement, but it has a
significant remaining problem that two threads *reading* the same
object will fight over a *write* lock for the same bucket. This change
creates a tracking object (ObjectUseData) that persists for the
lifetime of the object and is manipulated using atomic adds to track
the read and write counts. The object is looked up from a
concurrent_unordered_map, which only needs a read lock so two threads
each reading won't collide.

Change-Id: I49f429d93efa7f86ee8d8c4159dabaf6353ae22e