syncval: Init Sync Val Resource Hazard
Implement the resource state logic for reads/writes and pipeline
barriers.  Implement sufficient glue code to allow for a simple test of
the validity, state update, and barrier update operations.

The goal of this WIP change is to demonstrate the ability to track the
usage and barrier state of a specific resource (for example a range of
memory address, or an image subresource (or sub-resource range), within
application to a immediate record time state tracking for as the
resource state when a read-write-synchronization operations graph.

Specific review is needed on the implied theory of operation for the
chaining of dependencies (which is supported) and the extensiblity to
image layout transitions and queue family ownership changes.

Note that no attempt is made to implement the "logically earlier" and
"logically later" expansion of stage masks, nor the expansion of
meta-stages (such as GRAPHICS or ALL), as such expansions would be
performed on the *inputs* of the resource state update operations.

Includes support for tracking the affect of reads from multiple stages on
access state and hazard conditions.

Change-Id: I63b419ef1b8de5519091dcb4cf75ae282e9d3c7f