scripts: Clean up errors in stateless Python string list
In Python, two adjacent strings get concatenated implicitely.
Missing commas in multi-line string lists is a common source of bugs
causing unwanted string concatenation. In this case, it is clear that this
comma is missing by mistake and there should not be a concatenation.

Also removed duplicates from the manual validation list and regenerated
the parameter_validation.cpp file.

Change-Id: Ib83013300be2d159decffb22c183d43a77534030