corechecks: Fix ValidateImageBarrierAttachment resource search
The validation layer thinks that the renderpass in which the cmd
barrier is called doesn't have any references to image attachment
it is locking on, however, this is incorrect as the attachment is
inside of the pColorAttachment.

The function is searching for a reference in some places, but an
if/else statement is written so that if an extension is found,
all the other fields are not checked, even if the attachment is
not in the first examined field, and the validation layers
returns a false negative.

Change-Id: I4116ff3a647044ea0434af332568946c53e17c24