layers: More efficient cb_bindings to object_bindings linking
Previously this code did a hash table lookup (and insertion if not
present) followed by a second hash table lookup and insertion. Swap
the order of the checks and check the smaller table (cb_bindings)
first, and use a small_unordered_map for it. This makes there
usually be no hash lookup on the first table. Then replace the
second table with a vector, and just push_back instead of doing a
hash table insertion (each object can only be in this table once).
So now, the majority of the time, no hashing and no memory allocation.

Store a BASE_NODE pointer in VulkanTypedHandle, which can be used to
skip the hash table lookup if it's present.

Change-Id: Id6dcc03947268189762023aaf830a5f7880fa809