layers: Fix const violation in PreCallValidateQueueSubmit
The queryUpdate and eventUpdate lambda functions would write to the queue
and command buffer state, even though this method is meant to be const.
Fix this by having Validate and Record each call the functions with updates
being recorded into a local variable, which Validate ignores and Record
latches into the global state.