layers: Implement relaxed UPDATE_AFTER_BIND externsync rules in ThreadSafety
Make ThreadSafety track which descriptor sets allow UAB and model descriptor
updates to those sets as if they were reads. This is meant to allow concurrent
descriptor updates and concurrent update+bind, which the spec allows for UAB
bindings. This isn't perfect - really this is only supposed to be allowed for
specific bindings that are created with UAB, but we apply it to the whole set
to keep things simple.