Android: Make tests more reliable
Testing has observed occasional spurious test failures that
may be due to a race condition (if the device takes too long to
wake up, it may miss the HOME keycode or process it in the
wrong state).

These changes include a tentative fix for the above, and also
include changes that allow triage of a timeout failure.
(Currently, on a timeout failure, "adb" gets an error
"failed to stat remote object" when trying to pull the
executioin output, which causes the test script to exit
before logcat text can be collected.)

test_APK.sh:
- add "sleep 2" after every keyevent, to ensure that a keyevent is
  processed before the next keyevent is sent (and to better simulate
  human usage)
- don't wake up the device until after you have the device lock
- don't exit early on error - this exit prevents us from getting the
  logcat text that may be helpful in triaging a timeout failure
- re-establish the filtered logcat output - the full logcat is still
  available as an artifact, but only the filtered output is presented
  in the console, as was done historically
- minor bash syntax and spelling fixes