layers: Add range map view to image_layout
Add a "View" object which provides an iterable interface (in Subresource
space) to data stored in encoded range maps.

Replace the ImageLayoutMap::ConstIterator usage in the places where only
either current or initial maps are of interest.  Profiles as faster than
baseline (prior to range map) and faster than the parallel iterator
ConstIterator it replaces.

Change-Id: Ie53833b8c1e43dc1cec4c3b88a388e0fc48bd332