Fix Chromium warning about copying iterators.
"range-loop-constructs" is a new warning added to catch this bad code
pattern. Because the warning might be unavailable on older Clang or GCC
we don't try to enable it anywhere for now. If it pops up again we can
look at disabling it in the GN build or adding it to the default set
of enabled warnings.