travis: Improve ccache efficiency on Travis-CI builds
- Turn on CCACHE_COMPRESSION
- Enable ccache for GN builds

For some time, the ccache hit rate in Travis-CI builds has been
very low, often 0 hits per run.

This happens because there is a limit on the ccache size imposed
by Travis-CI and our builds are easily exceeding it, resulting in
severe cache thrashing.

In particular, glslang/spirv-tools are built first, filling up the
cache. Then the rest of the build objects push the glslang/spriv-tools
objects out of the cache, replacing them with their own.

In the next Travis-CI build, the retrieved cache does not contain the
glslang/spirv-tools objects, so they get recompiled and push
the other objects out, continuing the thrashing.

Turning on CCACHE_COMPRESSION lets the entire build fit into the
cache.  This is especially beneficial for the glslang/spriv-tools
code since it changes only when the known-good versions are updated.