scripts: Use named params for generator options
When calling in the vulkan header python generator the options have
changed in 1.2.142 which causes the options passed by the layers to be
off by one. This ends up setting the emitExtensions to the value of the
sortProcedure and eventually causes the script to fail.

This CL updates the layers to used named parameters for the generator so
that the addition of the genpath does not throw off the parameter
positioning.

Change-Id: I772bf6c651725314d6b39c9e9c61bd0dc5169fb4