layers: Fixed multiview handling in RenderPass code
The state_tracker and test code use the ConvertToRp2 functions
to copy all the original RenderPassCreateInfo information into
a RenderPassCreateInfo2 structure, but this was done incorrectly.
For instance, the pNext chain was simply copied, which is invalid.
Only VkRenderPassFragmentDensityMapCreateInfoEXT is valid for a
RPCI2 struct. This also affected the Multiview handling in the
Renderpass compatibility checks, as MV support in CreateRenderPass
requires the ext, while the data is in other places in
CreateRenderPass2.

Change-Id: Ic3af32ee4a10996ae50f298387dd676ccd47ea6e