layers: Fix depth/stencil resolve false positives
The depth/stencil resolve attachment in a render pass does not need to
have the same format as the depth/stencil attachment. The only
restriction is that if the resolve attachment has a depth or stencil
aspect, it must match in format and size with the depth/stencil
attachment.

This means the resolve attachment is free to remove image aspects (depth
or stencil) but not to add them. However, the layers were incorrectly
reporting errors when the resolve attachment did not have one of the two
aspects and the depth/stencil attachment did.

Fixes #2212