syncval: Add dst scope barriers to all ILT
Update the apply barriers logic s.t. the dst*Masks are applied for all
layout transitions.  Implicitly adds support for memory with no prior
access.  Does suppress *_AFTER_WRITE errors for memory access outside of
first scope but that will be caught as WRITE_AFTER_* errors for the
layout transition itself.

Fix test and BeginRenderPass validation that this change exposes issues
in (in this change for bisectability)

Change-Id: Ifa12ea9c234f550c1c68a62684809a54398f7358