syncval: Split tagging in VkCmdNextSubpass()
We need to consider the resolve and store portion of this
command to be part of the previous subpass, so it needs a
different tag from the layout transition and load part of
the command.

To do this, add a low order 'sub command' bit to
ResourceUsageTag.index, which lets the tags be different when
comparing. The layout the index field is now:

reset number[31 bits] | sequence number [32 bits] | sub command [1 bit]

Change-Id: I54cb435e2f6cfd467ab3dedf71b1091ee3fff736