[WIP] layers: Enable printf warnings for Log* methods
gcc and clang both can generate warnings about printf()
errors for any function if they are marked with:

__attribute__ ((format (printf, 4, 5)))

Obviously this change cannot submitted as-is, but it finds lots
of problems with our log messages.

The biggest downsides are that it doesn't support the stdint.h
format macros very well (PRIx64, etc..) and does it work on windows.

Change-Id: I91535aa12649f6a3a127f6141382a7598e198510