corechecks: Add validation of synchronization2 entry points
Refactor stage mask, access mask, and barrier validation to
be templated to handle the 3 new barrier types in thist s
extension. As part of this the many, many VUIDs for checking
these types are moved to tables in a separate file. This refactor
found a stage mask / access mask mismatch in an existing test.

Refactor queue submit verification to handle vkQueueSubmit2KHR(),
similarly. Unfortunately the the differences between the new
and old submit structures make this not work out quite as nicely.

Change-Id: Ia8f0133d34cf2d584fdd436aa2db643947ef91ba