build: Suppress -Wunguarded-availability-new on macOS in GN
On macOS vk_mem_alloc.h defines aligned_alloc and then uses it, which is
valid. However since aligned_alloc was added in macOS 10.15, clang
complains that it is used without it being in the deployment target (if
deploying for a version <10.15). This commit suppresses the warning at
the GN level since vk_mem_alloc.h is a copied from VMA.