layers: Sync syncval snippet tables with spec v1.2.169
This pulls in changes to ray tracing pipeline stages and access
masks.

VK_ACCESS_TRANSFORM_FEEDBACK_COUNTER_READ_BIT_EXT lost an allowed
pipeline stage in commit 5f585ae24. Note this is in a pending
change to the spec that is not yet in v1.2.169

A growing amount of manual fixups is required to pull this stuff
in to the script. I've noted them, but we really need to find
a better way to get these tables.

This fixes github issue #2542

Change-Id: Ia3b4dc2161babf5a9c3647a016303b5810ba5e15