scripts: Exit CI script if codegen check fails
Previously this was allowed to keep running and just reported the status
at the end. In hindsight, this just buries the error message in the CI
log.

Also change CI order to run codegen last so a failure doesn't interfere
with other checks. Finally, fix bug where failure code from LVT was not
passed back to the CI script.

Change-Id: I62a9a4c959258518a60a49d24b4ca0959ecc9cb9