layers: Handle nested transform feedback queries
Transform feedback queries are indexed for multiple streams.  That
means somthing like this is can be done:

vkCmdBeginQueryIndexedEXT(cmdBuf, xfbPool, queryIndex42, 0x0, 0);
vkCmdBeginQueryIndexedEXT(cmdBuf, xfbPool, queryIndex43, 0x0, 1);
vkCmdBeginQueryIndexedEXT(cmdBuf, xfbPool, queryIndex44, 0x0, 2);
...draw...
vkCmdEndQueryIndexedEXT(cmdBuf, xfbPool, queryIndex42, 0);
vkCmdEndQueryIndexedEXT(cmdBuf, xfbPool, queryIndex43, 1);
vkCmdEndQueryIndexedEXT(cmdBuf, xfbPool, queryIndex44, 2);

But as-is, the validation layer was generating an error on the second
begin-query saying that a query of that type was already active.

This fix is to also compare the query's index in addition to its type
during validation.  For non-xfb queries the index is always zero.