layers: Fix VUID-VkMappedMemoryRange-size-01389/90
For VUID-VkMappedMemoryRange-size-01389, the layers were incorrectly
applying the spec requirement that the mapping must end at a multiple of
the non-coherent atom size. Instead, they were calculating if the end of
the allocation minus the range offset was a multiple of the non-coherent
atom size.

In addition, this commit also takes into account the new possibility of
the memory mapping actually ending at the end of the allocation, which
is considered legal.

For VUID-VkMappedMemoryRange-size-01390, the error message is improved
to show every affected value and the second part of the spec condition.