layers: Add an Android default log force option via setprop
It's useful for certain downstream tools/systems(like automation)
to parse the log output of validation layers. On Android via logcat,
the default validation layer callback only prints when the Vulkan app
explicitly doesn't subscribe to the VK_EXT_debug_utils callback via
vkCreateDebugUtilsMessengerEXT. When the app does subscribe, then only
the app's callback is executed. To allow such downstream tools to
parse logcat for validation layer errors, this PR adds a mechanism to
force the default log to always be printed, regardless of whether the
app subscribes via vkCreateDebugUtilsMessengerEXT. This PR is currently
limited to Android-only, and opt-in with an Android system property.
This allows downstream tools to set the property and parse the logcat
in a known form(accessing key info like VUID), and every other existing
app relying on validation layers will remain unaffected.