scripts: Switch from mkpath to makedirs
distutils.dir_util.mkpath will not re-create a directory if it has been
removed with shutil.rmtree because it caches the filesystem under the
hood. This was causing the clone retry to fail on the second iteration
every time.

Switch to using os.makedirs for equivalent functionality without the
cache.

Change-Id: Ibd8169cff4573a5dba9900b10bcc60291b40e787