tests: Fix tests to run with ASan enabled
Note there still are many leaks reported with <unknown module> in the
stack but no ValidationLayers or test code. These might be caused by
failing to clean up vulkan objects allocated by the driver.
But the tests will run to completion and it is possible to see leaks
or other errors.