layers: Remove redundant handle members from state objects
Now that BASE_NODE has a typed handle member, replace handle
fields with methods in all derived state objects. This is done
as a separate step because it is very mechanical and repetitive.

It could be cleaned up even further with template magic in
vk_layer_logging.h and the chassis Log*() methods.