tests: Skip some tests that don't run on RADV
GpuBufferDeviceAddressOOB:
The shader in this test does not run correctly even without GPU-AV
A positive test written to exercise it got no validation errors, yet
the shader never executed correctly.  The same test passed on other
platforms.

Swapchain tests that bind swapchain memory to images:
As mentioned in the skipping of these tests on Android,
VUID-VkBindImageMemoryInfo-pNext-01631 specifies that
If the pNext chain includes a VkBindImageMemorySwapchainInfoKHR
structure, memory must be VK_NULL_HANDLE, but future calls
crash when this is followed.

Change-Id: I0a6081771977e848f65cae5a7fb4faad306f734b