layers: Improve VUID-VkBindImageMemoryInfo-memory-02629 checking
When the dedicatedAllocationImageAliasing is enabled, it should be
possible to bind a compatible image to a VkDeviceMemory even after
the original dedicated image is destroyed. Fix this check by caching
the dedicated create info in DEVICE_MEMORY_STATE and using this instead
of looking up the dedicated image at bind time.

Fixes #2730