layers: Move CommandBuffer state objects to separate files
This is the last of the deck chair rearranging for a while. As of this
change core_validation_types.h is no more. Instead, in state_tracker.h
there's some forward declarations in state and *minimal* includes of
other state object headers. This greatly reduces the changes of hitting
the 'change one small thing and rebuild the universe' problem.