scripts: Fix crash with dynamic rasterizer enable
The following CTS tests were crashing on NVIDIA when enabling validation
if GPU-assisted validation was enabled as part of it.

dEQP-VK.pipeline.extended_dynamic_state.*.enable_raster

The crash was happening in the driver, but the culprit appeared to be
the layers removing some structure pointers from the graphics pipeline
create info structure that was being passed down to the driver. These
pointers were being removed because rasterizerDiscardEnable was set to
true in the original structure. However, it was being set to false later
using VK_DYNAMIC_STATE_RASTERIZER_DISCARD_ENABLE_EXT and a call to
vkCmdSetRasterizerDiscardEnableEXT.

To fix the issue, make the layers take into account rasterization
discard can be a dynamic state and avoid removing some structures if
that's the case.