layers: Clean up of Surface and Swapchain state tracking
Swapchains should be the parent their Surface in the object
tracking tree because destroying the Surface will invalidate
the Swapchain. Make their state objects follow the convention
of 'parent holds shared_ptr on child object'.

Because Surfaces are tracked at the Instance level, Swapchains
need to be explicitly destroyed in DestroyDevice() to avoid
potential accesses to the Device level state after it is destroyed.
VkBestPracticesLayerTest.ValidateReturnCodes + ASan hits this
condition.