layers: Fix state recording of 2D ImageViews of 3D Images
Layout transitions of these ImageViews affect the entire mip level,
not just the slice the view references, so don't use the slice
aliasing in NormalizeSubresourceRange(). Add methods to retrieve
the depth slice offset and extent so that sync validation can
use it.

This fixes additional corner cases uncovered by the fix for Issue #2910.