layers: Clean up most of PIPELINE_STATE
The biggest part of this change is moving all of the 3 different
CreateInfo structures into a single union. Because the sType
fields of all these structures are at the same location, only
one field now needs to be checked to figure out what type
of pipeline is represented by the state object.  Another possible
clean up would be to make separate subclasses for each type of
pipeline.