layers: Clean up PipelineStageState creation
Parts of PIPELINE_STATE were set up during creation of PipelineStage
objects, in ValidationStateTracker::RecordPipelineShaderStage().
Refactor the constructors so that more fields in PIPELINE_STATE,
including all of the PipelineStageStage objects, can be const.
Note that the active_slots map is almost, but not quite const.
Since this map is used performance critical parts of DescriptorSet
state tracking and validation, it will probably need further work