layers: Fix min/maxTexelGatherOffset check for dynamic gather
The offset argument to OpImage*Gather is verified against
minTexelGatherOffset and maxTexelGatherOffset.  However, this offset is
not always constant.  The id of the offset was used to look up the
instruction that defines this "constant", which returned an end()
iterator for dynamic offsets.  Accessing this iterator is an OOB bug.