layers: Use concurrent_ordered_map for state objects
Use the thread safe unordered_map for all state objects. This makes
it safe to call Add<>(), Get<>() or Remove<>() without holding a
lock. Note that because Get<>() returns a shared_ptr to the state
object, it will not be deleted while the caller holds this reference.

This does make iterating all state objects even more expensive, but
that was already a bad thing to do anyways.