layers: Dehandle the image layout maps
Change the keys in the CMD_BUFFER_STATE and CoreChecks image layout
maps to be IMAGE_STATE* rather than VkImage. Layout validation
requires the IMAGE_STATE object so this saves many lookups in cases
where the entire maps are processed. The CMD_BUFFER_STATE layout
map can be kept up to date by removing entries that are being
destroyed in the NotifyInvalidate() callback.