layers: Encapsulate SEMAPHORE_STATE member data
Make all non-const member data private and restructure semaphore
tracking to allow it to eventually be made thread safe. The
biggest change is that each semaphore's state now keeps a
queue of pending operations, which removes the need to
frequently walk all pending submissions in every queue.

Fixes #3590, which was introduced when one the 'for every queue'
loops was removed in a previous commit.