layers: Make QUEUE_STATE and related classes threadsafe
It turns out the QUEUE_STATE deque does not need to be public, so
all of these classes are fully encapsulated and can manage their
locking.

The tricky part of this change is making sure QUEUE_STATE::Retire()
doesn't get called with a lock held. This would lead to deadlock.