layers: Make CMD_BUFFER_STATE threadsafe
CMD_BUFFER_STATE is much more complicated than other state
objects. At this point it is not possible to encapsulate
all of its non-const member data without massive code changes
and probably much worse performance.

Instead, code that uses this object must manage its locking.
The easiest way is to use GetRead() or GetWrite() to retrieve
a pre-locked pointer. Code in a class derived from
ValidationStateTracker MUST NOT hold a lock on a command buffer
when it calls a method in the base class or deadlock will result.

There are some holes in the locking. The lamda / std::function callbacks
used to set up submit-time validation are not correctly locked.
This is because for secondary command buffers, some of the callbacks
are transferred to the primary command buffer so it isn't easy to keep
track on if the state object is already locked, or not. There is also
a hole when both the state tracker and core validation have Record
methods that update members of CMD_BUFFER_STATE. See comments in
core_validation.h for the details. This will need to be improved in
later PRs to make it possible to use fine grained locking in the other
types of validation that use the state tracker.