build: Skip find_package for added targets
Amber uses add_subdirectory for some of the dependencies validation
layers searches using find_package. In case these are not installed
on the system the already added dependencies are not found. On the
other hand, if they are installed, validation layers would use those
instead of the locally checked out dependencies which should be
prioritized.