syncval: Handle destruction of bindless descriptor resources
If a descriptor set is 'bindless, it is valid to destroy
buffers or images (or views) used by one of its descriptors, as long
as that descriptor is not 'dynamically used' by any shader invocations.
Because these descriptor sets are often quite large, the state tracker
does not remove state pointers from descriptors when a resource is
destroyed. Avoid crashing in this case by checking for destroyed state
objects.

Fixes #3717