layers: Avoid out of bounds access in UpdateDrawState()
Binding a pipeline that uses n descriptor sets with only n - 1
of those descriptor sets bound with the validation layers from
the Vulkan SDK 1.3.204.0 gives a crash in STL code beneath
CMD_BUFFER_STATE::UpdateDrawState() line 1007.

Locally built validation layers in debug mode gives an assert
that access to state.per_set[set_index] in
CMD_BUFFER_STATE::UpdateDrawState() is out of bounds. Where
set_index is 5 but there are only 5 elements in state.per_set.
Note that assert is not the same position as the crash but it
seems to be the same call-stack.