layers: Add add'l support for surfaceless_query ext
This extension allows a VkSurface handle to be null for a handful
of query APIs. This is because the target devices will return the
same support for all surfaces. If the ext is enabled and the
surface handle is NULL, the queried surface data will be stored in
and retreived from the physical_device_state structure, instead of
a surface_state structure.

Change-Id: Ifb7d548a445d327bee7c18928a6b539228b2bf65