practices: Flag conflicting depth directions
Track depth function directions and flag images with conflicting ratios.

Z-cull depends on the direction of the depth comparison op.

This commit tracks the ratio at a subresource granularity and flags when
the ratio of draws using one or the other exceeds 20%.

For example, if 10% of the draws use LESS and 90% GREATER, nothing gets
flagged. But if 30% is used for LESS, then it will flag a message.

The Z-cull direction can be reset discarding the contents of the image,
or clearing it.