layers: Fixes memory leak in descriptor set validation
Fixes memory leak in descriptor set validation. On allocation of new
descriptor sets, the sets are added to a map that they are never
removed from on pool reset, pool destruction, or set free.  This
causes a memory leak to manifest itself when resetting the pool and
reallocating the descriptor sets each frame.