layers: Fix Dynamic Rendering SEGFAULT for Portability
When validating graphics pipeline creation for portability
implementations, no check was implemented to determine if the pipeline
was using dynamic rendering and would have a null render pass, causing
a SEGFAULT by attempting to dereference a null pointer to the render
pass state. A check for whether the VkPipelineRenderingCreateInfo
struct was included in the pNext chain of the
VkGraphicsPipelineCreateInfo struct was implemented to fix this.