layers: GetLayerEnvVar shouldn't change the layer_config object
This function was using layer_config.vk_layer_disables_env_var
as temporary storage so that it can return a char * which doesn't
have permanent storage. "Fix" this by using a static std::string
instead.

But really, there's not a safe way to implement this function on
all platforms we support. GetEnvironment() returns a std::string,
which solves this problem. Use that instead.

Fixes #3048