layers: Remove Descriptor type and class members
With the exception of mutable, the class and type is the same
for an entire binding, so we can avoid storing them repeatedly.

All of the type changing stuff moves down into MutableDescriptor.
It could be cleaned up much more, using unions so that it doesn't
waste as much space.