build: Add GitHub Actions dependency caching
Cache the builds of components that this project depends on to speed up
GitHub Actions builds.

The dependent components are specified in scripts/known_good.json whose
hash is used as a key for the cache.  When the known good file is
updated, a cache won't be found for that hash/key and the components are
rebuilt and cached again.

For the more common case for when the known good file does not change,
the cache is found and the dependent components don't need to be
rebuilt.

Documentation for this GitHub Actions feature can be found at:
https://docs.github.com/en/actions/using-workflows
in the "Caching dependencies to speed up workflows" section.  There are
also APIs available for managing the caches.