build: Improve GitHub Actions caching
- Extend update_deps.py to skip repos if install dir is present
- Cache only build/install directories from dependent repos in the
  external directory

This change keeps update_deps from fetching a repo on top of a cached
repo and building an already built repo. (Previously, the entire
external directory was cached.)

It also significantly lowers the amount of data cached, allowing more
caches (related to versions of known_good.json) to be saved.