layers: Use a thread for QUEUE_STATE submission processing
Rewrite QUEUE_STATE to use a worker thread to update state. This was
necessary to fix a race condtion if multiple threads wait on the
same timeline semaphore.

Thread 1:
vkWaitSemaphores() : wait on semaphore for value N

Thread 2:
vkQueueSubmit()    : signal semaphore for value N
vkWaitSemaphores() : wait on semaphore for value N

Thread 1 starts waiting before we have any information about which
queues(s) will need to make progress before value N will be complete.
In the old implementation, that might cause it to not correctly update
the queue(s) state after it woke up. That would then lead to false
'object in use' errors. In theory, it might have been possible to fix
this problem, but the logic was already confusing. Also, this opens up
the possiblity of moving execute time validation and state updates to
the queue thread, which might improve performance.

Fixes #4308