layers: Fix CB locking during query callback functions
There were deadlocks and race conditions when running the callbacks
in CMD_BUFFER_STATE::queryUpdates. This was because secondary
command buffer callbacks were copied into the primary's callback
list. Then the callbacks were all called with the primary buffer's
lock locked, but not the secondary. Instead, make a callback that
loops through the secondary's callbacks in place rather than
copying. Then the callback can safely lock the secondary.

Hopefully this is only an interim fix and we can find a better
way to implement this without callback lambda voodoo.

Fixes #4243