layers: Save multisample state if render pass is not null
Several tests in the dEQP-VK.transform_feedback.* group are passing a
valid pMultisampleState pointer when creating pipelines with
rasterization disabled, and the layers are still generating
VUID-VkGraphicsPipelineCreateInfo-renderpass-06631.

The VUID requires multisample state to be passed if the pipeline
contains fragment shader state and a non-NULL render pass, but the
layers do not save the provided multisample state info due to
rasterization being disabled and later emit the validation error
incorrectly.

This commit makes the layers save the multisample state if a render pass
is provided.