Update WAR for NV_SCI issue
VulkanSC-Headers mistakenly has VK_OBJECT_TYPE_SEMAPHORE_SCI_SYNC_POOL_NV
ifdef'd with VK_USE_PLATFORM_SCI, unlike other VkObjectType's corresponding
to other extensions. As a result a work around is needed to deal with the
use of this VkObjectType in generated vksc layers code. Previously this was
done by turning on NV_SCI support by default, but this can cause issues on
platforms without the proper nvsci headers. This change updates the vksc
layers such that the use of the VkObjectType is also ifdef'd as a
temporary work around until the headers are fixed.