tests: Remove TestMultipleQueuesWaitingOnSemaphore
This test did:

 vkQueueSubmit(queue_a, [signal sem])
 vkQueueSubmit(queue_a, [wait sem, signal sem])
 vkQueueSubmit(queue_b, [wait sem])

It expected VUID-vkQueueSubmit-pWaitSemaphores-00068 to fail.
However, that says:
  When a semaphore wait operation referring to a binary semaphore
  defined by any element of the pWaitSemaphores member of any element
  of pSubmits ***executes*** on queue, there must be no other
  queues waiting on the same semaphore.

But [signal sem], [wait sem, signal sem] on queue_a will have completed
before the submission on queue_b executes, so this was not an error.
The only way to trigger this VUID would be:
 vkQueueSubmit(queue_a, [signal sem])
 vkQueueSubmit(queue_a, [wait sem])
 vkQueueSubmit(queue_b, [wait sem])

That case is covered by QueueSubmitWaitingSameSemaphore.