state: Simplify CMD_BUFFER_STATE object tracking
Remove the `framebuffers` map since it is redundant with
`object_bindings`.

Work around a possible race condition in NotifyInvalidate. Due to
interactions between the BASE_NODE::tree_lock_ and
CMD_BUFFER_STATE::lock it was possible for a (not really) 'bad'
binding to get added to `broken_bindings`. Try to fix this by
only adding a node to `broken_bindings` if it was still found in
`object_bindings`.