layers: Clean up vk_layer_logging.h
Add vk_layer_logging.cpp, so that most of the logging functions
can be de-inlined. This appears to reduce compiled code size for
the validation layer by 11% to 33% (on linux).

Change LogMsgLocked() to LogMsg() and restructure it to avoid
duplicated code in class ValidationObject.

Remove use of vasprintf() and snprintf() into a std::string instead.
vasprintf() calls malloc internally and VVL must call free. When
overriding malloc, this gets complicated because libc and VVL can
end up using different malloc and free implementations.