tests: Fix jump to null function pointer
Found on lavapipe 22.3.2.

Lavapipe doesn't support the VK_KHR_SAMPLER_YCBCR_CONVERSION_EXTENSION.

This test uses this extension by calling one of its functions to test
the case where one of its dependent extensions isn't enabled.  The
trouble with this is that if the sampler extension isn't supported, the
test should not call the extension function and should skip the rest of
the test.

This commit modifies the check for the extension to skip the test in
this situation.