chassis: Fix memory leak in vkCreateInstance if an error occured
If an error occurs duing the call down the CreateInstance chain, VVL must
clean up any allocated objects. This is made easier by only allocating
Validation Objects for which the filters allow, rather than the previous
method of allocating everything and removing the unused ones. While from a
first glance std::unique_ptr<> seems like the obvious choice, because of
how intrusive such a change would be it was not chosen.

The second place an error can occur is during PostCallRecordCreateInstance.
This results in faileding a vulkan function call after a successful call
down the vkCreateInstance chain, leaving Vulkan in a weird state. The
previous commit would not pass the error code upwards, so to the application
everything appeared correct since the return value was VK_SUCCESS. However,
this causes CTS Out Of Memory tests to fail due to OOM not being returned.
The fix for this is to make the PostCallRecordCreateInstance pass the
VkResult value by reference instead of value, allowing the error code to
propagate upwards. This did require special casing the PostCallRecord
in the various code generators.

Lastly, the commit does a general cleanup of the ValidationObject code by
removing the functions only used during initialization. Since they are used
only during vkCreateInstance and vkCreateDevice, moving them into those
functions cleans up the overal interface of the functions.

This commit makes the following tests have clean output when run with a
leak sanitizer:
* VkLayerTest.InstanceExtensionDependencies
* VkLayerTest.InstanceBadStype
* VkLayerTest.InstanceDuplicatePnextStype
* VkLayerTest.InstanceAppInfoBadStype
* VkLayerTest.InstanceValidationFeaturesBadFlags
* VkLayerTest.InstanceBadValidationFlags

Note that to get the clean output, the loader must have been updated to
contain the commit 3db19f3 (or just run with a loader built against
1.3.243 or higher).